- Work around DWM compositor screwing up when duplicating frames by adding "DWM triple buffering" option (on by default) which sets presentparameters with frame queueing for all blitters. - Work around DWM breaking windows with frame queuing after composition is disabled and re-enabled, by resetting present parameters on next video frame after composition enable event. - Work around DWM breaking OpenGL window on minimize by unsetting present params on hide event, and setting on show event. - Ignore sync to vblank when DWM composition is active. - Deactivate sync to frame rate option when DWM composition is active. - Make window size setting a menu rather than a part of video settings dialog. - Work around Windows file dialog blocking events by waiting for pause to complete before launching such dialog. - Work around events getting blocked when resizing on Windows by consuming pending blit requests on resizeEvent. - WASAPI: Detect and use shared stream sampling rate in shared mode because all other rates appear to fail. - Default screen names start at "1" rather than "0". - Build fixes for mingw gcc 3.4. - Refactoring.
git-svn-id: https://gambatte.svn.sourceforge.net/svnroot/gambatte@259 9dfb2916-2d38-0410-aef4-c5fe6c9ffc24

+++ common/resample/src/chainresampler.cpp
@@ -77,7 +77,7 @@ void ChainResampler::adjustRate(const long inRate, const long outRate) {
 	
 	exactRatio(mul, div);
 	
-	bigSinc->adjustDiv(static_cast<double>(inRate) * mul / (static_cast<double>(div / bigSinc->div()) * outRate) + 0.5);
+	bigSinc->adjustDiv(static_cast<int>(static_cast<double>(inRate) * mul / (static_cast<double>(div / bigSinc->div()) * outRate) + 0.5));
 	
 	reallocateBuffer();
 	setRate(inRate, outRate);
